Change spine-cpp folder name
