+++ .gitignore
@@ -363,3 +363,5 @@ MigrationBackup/
 FodyWeavers.xsd
 /MonmusuPlayer/deps/SFML-2.6.1
 /MonmusuPlayer/deps/spine-cpp-3.8
+/MonmusuPlayer/deps/nlohmann
+/MonmusuPlayer/deps/spine-cpp
